Add Revlink for Workflow Runs from Workflow. Restructure Workflow.js into directory 'Workflow' w/ an index.js.
