+++ src/encoded/static/components/viz/Workflow/index.js
@@ -1,0 +1,6 @@
+'use strict';
+
+import Graph from './Graph';
+export { parseAnalysisSteps, parseBasicIOAnalysisSteps } from './parsing-functions';
+
+export default Graph;
