+++ src/encoded/types/workflow.py
@@ -26,6 +26,22 @@ class Workflow(Item):
                 'arguments',
                 'analysis_steps',
                 'arguments.argument_mapping']
+    rev = {
+        'workflow_runs': ('WorkflowRun', 'workflow'),
+    }
+
+    @calculated_property(schema={
+        "title": "Workflow Runs",
+        "description": "All runs of this workflow definition.",
+        "type": "array",
+        "items": {
+            "title": "Workflow Run",
+            "type": ["string", "object"],
+            "linkTo": "WorkflowRun"
+        }
+    })
+    def workflow_runs(self, request):
+        return self.rev_link_atids(request, "workflow_runs")
 
 
     @calculated_property(schema={
